feat: enhance watchlist functionality and add sell stock request DTO
- Updated Watchlist.vue to support multiple watchlists, including creation, deletion, and item management.
- Implemented sorting and filtering for watchlist items.
- Added modals for creating watchlists and adding stocks to watchlists.
- Introduced a new SellStockRequest DTO for handling stock sale requests with sale price and date.

+++ backend/src/main/java/com/capitalot/dto/PortfolioPerformanceDto.java
@@ -5,6 +5,7 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
@@ -13,6 +14,6 @@ import java.time.LocalDateTime;
 @AllArgsConstructor
 public class PortfolioPerformanceDto {
     private LocalDateTime timestamp;
-    private Double totalValue;
-    private Double gainLoss;
+    private BigDecimal totalValue;
+    private BigDecimal gainLoss;
 }
